requirements.txt
bokeh==2.3.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==1.1.0
flask-http-response==0.0.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
